requirements.txt
pymongo>=4.7.0
motor>=3.3.0
sentence-transformers>=3.0.0
dnspython>=2.3.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
py-eureka-client>=0.11.1